backend/requirements.txt
Django==3.1.12
asgiref==3.8.1
django-cors-headers==3.7.0
djangorestframework==3.15.1
djongo==1.3.7
dnspython==2.7.0
pymongo==3.11.4
sqlparse==0.2.4
tzdata==2025.2
